Extract podcast item as a component
