+++ modules/features/search/src/main/java/au/com/shiftyjelly/pocketcasts/search/adapter/SearchPodcastViewHolder.kt
@@ -1,17 +1,21 @@
 package au.com.shiftyjelly.pocketcasts.search.adapter
 
+import androidx.compose.foundation.background
+import androidx.compose.material.MaterialTheme
+import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.ComposeView
 import androidx.compose.ui.platform.ViewCompositionStrategy
 import androidx.recyclerview.widget.RecyclerView
 import au.com.shiftyjelly.pocketcasts.compose.AppTheme
+import au.com.shiftyjelly.pocketcasts.compose.components.PodcastItem
+import au.com.shiftyjelly.pocketcasts.compose.theme
 import au.com.shiftyjelly.pocketcasts.models.entity.Podcast
-import au.com.shiftyjelly.pocketcasts.search.component.SearchPodcastRow
 import au.com.shiftyjelly.pocketcasts.ui.theme.Theme
 
 class SearchPodcastViewHolder(
     val composeView: ComposeView,
     val theme: Theme,
-    val onPodcastClick: (Podcast) -> Unit
+    val onPodcastClick: (Podcast) -> Unit,
 ) : RecyclerView.ViewHolder(composeView) {
 
     init {
@@ -23,7 +27,13 @@ class SearchPodcastViewHolder(
     fun bind(podcast: Podcast) {
         composeView.setContent {
             AppTheme(theme.activeTheme) {
-                SearchPodcastRow(podcast = podcast, subscribed = podcast.isSubscribed, onClick = { onPodcastClick(podcast) })
+                PodcastItem(
+                    podcast = podcast,
+                    subscribed = podcast.isSubscribed,
+                    onClick = { onPodcastClick(podcast) },
+                    modifier = Modifier
+                        .background(color = MaterialTheme.theme.colors.primaryUi01)
+                )
             }
         }
     }
