+++ modules/services/compose/src/main/java/au/com/shiftyjelly/pocketcasts/compose/components/PodcastItem.kt
@@ -1,6 +1,5 @@
-package au.com.shiftyjelly.pocketcasts.search.component
+package au.com.shiftyjelly.pocketcasts.compose.components
 
-import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
@@ -16,23 +15,24 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
-import au.com.shiftyjelly.pocketcasts.compose.components.HorizontalDivider
-import au.com.shiftyjelly.pocketcasts.compose.components.PodcastImage
-import au.com.shiftyjelly.pocketcasts.compose.components.TextP40
-import au.com.shiftyjelly.pocketcasts.compose.components.TextP50
 import au.com.shiftyjelly.pocketcasts.compose.theme
 import au.com.shiftyjelly.pocketcasts.models.entity.Podcast
 import au.com.shiftyjelly.pocketcasts.images.R as IR
 import au.com.shiftyjelly.pocketcasts.localization.R as LR
 
 @Composable
-fun SearchPodcastRow(podcast: Podcast, subscribed: Boolean, onClick: (() -> Unit)?, modifier: Modifier = Modifier) {
+fun PodcastItem(
+    podcast: Podcast,
+    onClick: (() -> Unit)?,
+    modifier: Modifier = Modifier,
+    subscribed: Boolean = false,
+    showSubscribed: Boolean = false,
+) {
     Column {
         Row(
             verticalAlignment = Alignment.CenterVertically,
             modifier = modifier
                 .fillMaxWidth()
-                .background(MaterialTheme.theme.colors.primaryUi01)
                 .padding(horizontal = 16.dp)
                 .then(if (onClick == null) Modifier else Modifier.clickable { onClick() })
         ) {
@@ -57,7 +57,7 @@ fun SearchPodcastRow(podcast: Podcast, subscribed: Boolean, onClick: (() -> Unit
                     color = MaterialTheme.theme.colors.primaryText02
                 )
             }
-            if (subscribed) {
+            if (subscribed && showSubscribed) {
                 Icon(
                     painter = painterResource(IR.drawable.ic_tick),
                     contentDescription = stringResource(LR.string.podcast_subscribed),
